refactor: move task comment to separate component
