+++ components/profile/ProfileUserInfo/ProfileUserInfo.module.scss
@@ -65,39 +65,6 @@
   }
 }
 
-.ghLink {
-  display: flex;
-  align-items: center;
-  padding: 8px;
-  background-color: colors.$purpleSecondary;
-  border-radius: 4px;
-  margin-right: 16px;
-
-  @media (max-width: media.$tabletS) {
-    margin-right: 0;
-    margin-top: 8px;
-  }
-}
-
-.login {
-  text-transform: lowercase;
-  font-size: 12px;
-  line-height: 16px;
-  font-weight: fonts.$weightMedium;
-  color: colors.$purplePrimary;
-}
-
-.ghIcon {
-  display: flex;
-  align-items: center;
-  margin-right: 8px;
-  svg {
-    width: 16px;
-    height: 16px;
-    fill: colors.$purplePrimary;
-  }
-}
-
 .bio {
   margin-top: 4px;
   font-weight: fonts.$weightRegular;
