+++ components/tasks/TaskComment/TaskComment.module.scss
@@ -1,0 +1,102 @@
+@use '/lib/styles/variables/colors.module' as colors;
+@use '/lib/styles/variables/fonts.module' as fonts;
+
+.commentsListItem {
+  display: flex;
+  flex-direction: column;
+  padding: 8px 16px 16px;
+  border: 1px solid colors.$strokeMain;
+  border-radius: 8px;
+  margin-bottom: 16px;
+}
+
+.commentHeader {
+  display: flex;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+.teacher {
+  display: flex;
+  align-items: center;
+  margin-top: 8px;
+}
+
+.teacherImgWrapper {
+  width: 32px;
+  height: 32px;
+  margin-right: 10px;
+}
+
+.teacherImg {
+  border-radius: 64px;
+}
+
+.teacherName {
+  font-weight: fonts.$weightMedium;
+  font-size: 12px;
+  line-height: 16px;
+  color: colors.$fontPrimary;
+  margin-right: 10px;
+}
+
+.time {
+  font-weight: fonts.$weightMedium;
+  font-size: 12px;
+  line-height: 16px;
+  color: colors.$fontSecondary;
+  padding-right: 24px;
+  margin-right: auto;
+  margin-top: 8px;
+  position: relative;
+
+  &::before {
+    content: '';
+    position: absolute;
+    left: -12px;
+    top: 50%;
+    transform: translateY(-50%);
+    width: 4px;
+    height: 4px;
+    border-radius: 64px;
+    background-color: colors.$purplePrimary;
+  }
+}
+
+.ghLink {
+  margin-right: 24px;
+}
+
+.attemptBadges {
+  display: flex;
+  align-items: center;
+  flex-wrap: wrap;
+
+  & > * {
+    margin-top: 8px;
+  }
+}
+
+.underAssessmentLink {
+  padding: 10px 16px;
+  font-weight: fonts.$weightBold;
+  font-size: 12px;
+  line-height: 12px;
+  color: colors.$statusFontBlue;
+  background-color: colors.$statusBgBlue;
+  border-radius: 64px;
+  text-align: center;
+  margin-right: 16px;
+
+  &:last-child {
+    margin-right: 0;
+  }
+
+  &:hover {
+    opacity: 0.8;
+  }
+}
+
+.contentWrapper {
+  margin-top: 12px;
+}
