+++ utils/attemptsToComments.ts
@@ -9,6 +9,7 @@ export function attemptToComments(attempt: ISingleAttempt): Comment[] {
             name: attempt.teacher.user.name,
             surname: attempt.teacher.user.surname,
             image: attempt.teacher.user.image,
+            login: attempt.teacher.profile.login,
           },
           attempt_score: attempt.score,
           content: attempt.comment,
@@ -28,6 +29,7 @@ export function attemptsToComments(attempts: IAttempt[]): Comment[] {
         name: attempt.teacher.user.name,
         surname: attempt.teacher.user.surname,
         image: attempt.teacher.user.image,
+        login: attempt.teacher.profile.login,
       },
       attempt_id: attempt.attempt_id,
       attempt_number: attempt.attempt_number,
